check_pkg {yulab.utils}	R Documentation

check_pkg

Description

Check whether the input packages are installed

Usage

check_pkg(pkg, reason = NULL, ...)

Arguments

pkg
package names

reason
the reason to check the pkg. If NULL, it will set the reason to the parent call.

...
additional parameters that passed to rlang::check_installed()

Details

This function check whether the input packages are installed. If not, it asks the user whether to install the missing packages.

Value

see also check_installed

Author(s)

Guangchuang Yu

[Package yulab.utils version 0.1.4 Index]