install_zip_gh {yulab.utils}	R Documentation

install_zip_gh

Description

install github package

Usage

install_zip_gh(repo, ref = "HEAD", args = "--no-build-vignettes")

Arguments

repo
github repo

ref
github branch, default is HEAD, which means the default branch of the GitHub repo

args
argument to build package

Details

it download the zip file first and use install_zip to install it

Value

No return value, called for installing github package

Author(s)

Guangchuang Yu

[Package yulab.utils version 0.1.4 Index]